chore: Rewrite data.py file into separated files into data package
