+++ src/squarecloud/data/base.py
@@ -1,0 +1,21 @@
+from __future__ import annotations
+
+from typing import Any
+
+from .._internal.constants import USING_PYDANTIC
+
+if USING_PYDANTIC:
+    from pydantic.dataclasses import dataclass
+else:
+    from dataclasses import dataclass
+
+
+class DataClasMeta(type):
+    def __new__(cls, name: str, bases: tuple, dct: dict[str, Any]) -> type:
+        new_class = super().__new__(cls, name, bases, dct)
+        return dataclass(frozen=True)(new_class)
+
+
+class BaseDataClass(metaclass=DataClasMeta):
+    def to_dict(self) -> dict[str, str | dict[str, Any]]:
+        return self.__dict__.copy()
