+++ src/squarecloud/data/base_database.py
@@ -1,0 +1,8 @@
+from .base import BaseDataClass
+
+
+class BaseDatabaseData(BaseDataClass):
+    id: str
+    name: str
+    type: str
+    cluster: str
