+++ src/squarecloud/data/dns_record.py
@@ -1,0 +1,8 @@
+from .base import BaseDataClass
+
+
+class DNSRecord(BaseDataClass):
+    type: str
+    name: str
+    value: str
+    status: str
