+++ src/squarecloud/data/language.py
@@ -1,0 +1,6 @@
+from .base import BaseDataClass
+
+
+class Language(BaseDataClass):
+    name: str
+    version: str
